requirements.txt
aiodns==3.0.0
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.4
async-timeout==4.0.2
attrs==22.2.0
banal==1.0.6
Brotli==1.0.9
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
ciso8601==2.3.0
colorlog==6.7.0
dataset==1.6.0
easygoogletranslate==0.0.4
frozenlist==1.3.3
hikari==2.0.0.dev116
hikari-lightbulb==2.3.1
hikari-miru==3.0.2
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.2
multidict==6.0.4
orjson==3.8.12
pycares==4.3.0
pycparser==2.21
python-dotenv==0.21.1
requests==2.30.0
SQLAlchemy==1.4.48
typing_extensions==4.5.0
urllib3==2.0.2
yarl==1.9.2
